[CornerstoneHealthcareUS] Add category and use DictParser
